Store the presentational state of Reading Mode webui in ReadAnythingController
The presentation_state_ tracks who is the current owner of the WebUi.
When it's handed off to the side panel, the presentation_state_ is
updated to kInSidePanel. (When Immersive Mode is implemented and the
WebUI is handed off to the Immersive view, presentation_state_ will be
updated to kInImmersiveOverlay.) When the WebUI is handed back off to
the ReadAnythingController, it is updated to be kInactive, because the
WebUI is not being displayed.

Bug: 450948683
Change-Id: I1a4d954697b097f985957b99d849170d6ca0ffbb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7254455
Reviewed-by: Kristi Saney <kristislee@google.com>
Commit-Queue: Eitan Goldberger <eitang@google.com>
Cr-Commit-Position: refs/heads/main@{#1557803}