[lensoverlay] Add accept/dismiss buttons to non-blocking privacy notice.
Focusing the searchbox dismisses the notice and is no longer treated as
accepting the notice.

If the user performs an interaction without clicking the accept button
in the notice, this is treated as granting permission for the current
session only. The notice will continue to appear on subsequent
invocations until the accept button is clicked.

Suggestions in the searchbox are disabled until the user accepts the
notice, after which the searchbox will be focused and suggestions will
be fetched.

Bug: 467451572
Change-Id: I1a5c96b48289868caf55d3837184db71ca3d7576
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7248304
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Reviewed-by: Mustafa Emre Acer <meacer@chromium.org>
Commit-Queue: Bryan Nguyen <nguyenbryan@google.com>
Reviewed-by: Duncan Mercer <mercerd@google.com>
Cr-Commit-Position: refs/heads/main@{#1557814}