[Desktop android] Implement MoveGroupTo 2/2
Implement moving to the closest valid index. This is only needed if the
index to move to is in the middle of another tab group. In which case
the closest valid index should be just before or just after that
tab group. "Ties" (when there would be 2 tabs before and 2 after
from the index of the group) are deterministic.

Bug: 460650221
Change-Id: If58566b0723c5a9d2ec8ca0a81ab6be9ecc1a0ea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7241117
Reviewed-by: Alison Gale <agale@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557833}