[webui omnibox] Notify keyword changes only when entering or exiting
Problem: The OmniboxEditModel previously notified observers for any
change in keyword selection, including hint toggles and string updates.
This led to redundant notifications when the keyword state had not
meaningfully changed e.g. [1].

Solution: Restrict notifications to when a keyword is added or removed.
Before, a notification was also sent whenever changes were made to the
keyword hint. This is preferred because it lowers the amount of signals
sent to the popup for hiding the context menu button, making sure it
hides when we enter a keyword mode and hides when we exit one.

demos-
* Before: http://screencast/cast/NjYwMDQ3ODk4ODE3MzMxMnwwMDRjMmI2Zi02Mg
* After: http://screencast/cast/NDczMDQ5MDU1OTkyMjE3NnwyNTA1Y2VjMi1kNA

[1]https://screenshot/5CjgXQriAjAsupL

Change-Id: Ic62eb43a597d8605ffce6bb75bdee2955e2d9904
Fixed: 468091168
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7254097
Reviewed-by: Moe Ahmadi <mahmadi@chromium.org>
Commit-Queue: Moe Ahmadi <mahmadi@chromium.org>
Auto-Submit: Paul Adedeji <pauladedeji@google.com>
Cr-Commit-Position: refs/heads/main@{#1557834}