[syntheticresponse] Do not set special CSPs in dry-run mode
The dry-run mode in synthetic response installs the service worker, but
it actually use the synthetic response. So it's fine not to be set the
special CSP policy because we don't modify the response at all when the
feature is running as dry-run mode.

Bug: 466268411
Change-Id: I97d3a36a6bb960a3724fc914a8a2b36f2b2ec6d8
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7253071
Reviewed-by: Yoshisato Yanagisawa <yyanagisawa@chromium.org>
Commit-Queue: Shunya Shishido <sisidovski@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557835}