[Extensions] Fix TypeError in DragAndDropHandler on chrome://extensions.
Fixes a JavaScript error in DragAndDropHandler where
DataTransferItem.webkitGetAsEntry() could return null if the dropped
item is not a file or is not in read/write mode. This change adds a null
check before accessing the isDirectory property.

This also adds tests for the null scenario and for other drag and drop
scenarios.

Bug: 466050994
Change-Id: I9af3f31033ed027663c728c6decf73852c742e22
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7248839
Reviewed-by: Kelvin Jiang <kelvinjiang@chromium.org>
Commit-Queue: Kelvin Jiang <kelvinjiang@chromium.org>
Auto-Submit: Justin Lulejian <jlulejian@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557837}