[gardener] Update expected file for auto-attach-wait-for-debugger-2-sessions-worker
The test is failing on mac, caused by https://crrev.com/c/7248977 which
removed insignificant trailing spaces from the output of
testRunner.log() and changed the log output format (inserting a space
between arguments).
https://ci.chromium.org/ui/test/chromium/%3A%2F%2F%5C%3Ablink_web_tests!webtest%3A%3Ahttp%2Ftests%2Finspector-protocol%2Ftarget%23auto-attach-wait-for-debugger-2-sessions-worker.js

This CL removes a space from the expected file for mac-14, and remove
the expected file for mac-13 because it's the same as the common
expected file, so it's no longer necessary.
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/web_tests/http/tests/inspector-protocol/target/auto-attach-wait-for-debugger-2-sessions-worker-expected.txt

Change-Id: I66560663191a24e52d1f80e9932fe7bc58a5610f
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7255574
Commit-Queue: Anna Sato <annasato@chromium.org>
Reviewed-by: Takashi Nakayama <tnak@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557861}