Specifies a null TimeTicks for input_start time when unsure
The input_start field must be set for the navigation when there is a
user input event. The value must be the OS-level timestamp. But the
current `NavigationRequest::CreateForSynchronousRendererCommit()`
specifies the current time instead of the timestamp of the user input
event. This CL specifies a null TimeTicks instead.

Ref:
https://source.chromium.org/chromium/chromium/src/+/main:third_party/blink/public/mojom/navigation/navigation_params.mojom;l=308-311;drc=c095fc3a4c773ca8dbf348fb375daefa28b9a816

https://docs.google.com/document/d/1Vk6ZyQcTrEUmUCEKU-NE0aDYOS0WGmGHE6jmph__DcI/edit?tab=t.0#heading=h.esot33hckmf

Bug: 444575624
Bug: 467158532
Change-Id: I7fd0feab9b565ccd8d260a4532e8ee4355510955
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7250968
Reviewed-by: Rakina Zata Amni <rakina@chromium.org>
Commit-Queue: Minoru Chikamune <chikamune@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557890}