Snapshot UKM `source_id` early into `PermissionRequest`
This ensures that the correct `source_id` is used for recording a
permission request `IGNORED` action, even if that "ignoring" happens
implicitly as a result of the user navigating away from a tab while a
permission prompt is showing (or a permission request still being
queued). In these two cases, getting the `source_id` when the new
navigation commits is too late, as it already corresponds to the newly
committed page; but prematurely treating the request as "ignored" as
soon as the navigation provisionally starts would be too early (as it
may still fail).

Change-Id: Idca928b176900d63881b44950b626fec982d64d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7238754
Commit-Queue: Balazs Engedy <engedy@chromium.org>
Reviewed-by: Antonio Sartori <antoniosartori@chromium.org>
Auto-Submit: Enrico Bacis <enricobacis@google.com>
Reviewed-by: Balazs Engedy <engedy@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1557948}