Rename blur-value accessors on ShadowData
Rename Blur{,XY}() to BlurValue{,XY}(). These are "raw" accessors that
can be used when semantics are not important - or as an escape hatch
when the semantic accessors could end being confusing (like for the
drop-shadow() filter function).

Add BlurRadius() and BlurAsSigma(), where the latter converts a stored
radius value to a sigma value.

Store the values using a gfx::SizeF rather than a gfx::PointF. Return a
const StyleColor& from the GetColor() to avoid copies.

This is preparation for moving BlurRadiusToStdDev() to ShadowData.

Bug: 399058536
Change-Id: Ie22d4471518b37c2d990a0913a0ee509d81eece6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7252849
Commit-Queue: Fredrik Söderquist <fs@opera.com>
Reviewed-by: Florin Malita <fmalita@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558004}