[AutofillAi - M4] Create java version of an EntityInstance
In order to eventually add and edit entities from clank, we need a java
representation of an entity, which will be forwarded to c++ to interact
with the EntityDataManager.

This is an initial boilerplate, this class will likely be
significantly extended.

Bug: 411324196
Change-Id: Ic7819923e06d4f0410cf60253202003ccb6d0aea
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7176329
Reviewed-by: Dominic Battre <battre@chromium.org>
Auto-Submit: Bruno Braga <brunobraga@google.com>
Reviewed-by: Timofey Chudakov <tchudakov@google.com>
Commit-Queue: Bruno Braga <brunobraga@google.com>
Cr-Commit-Position: refs/heads/main@{#1558015}