[wv][metrics] Log foreground/background state for unhandled WebView renderer terminations.
This change updates the histogram
"Android.WebView.OnRenderProcessGoneResult" to track whether the app was
in the foreground or background before we kill/crash it due to unhandled
ORPG callback.

Note that the foreground/background state tracked here is different from
the foreground/background state of WebView that we typically track.
Those track the state of the WebViews in an app while this tracks the
state of the app as a whole.

Bug: 465703913
Change-Id: I01841033827345d2ccbb55cc64e073fe5513de9e
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7247849
Reviewed-by: Richard (Torne) Coles <torne@chromium.org>
Commit-Queue: Chidera Olibie <colibie@google.com>
Auto-Submit: Chidera Olibie <colibie@google.com>
Reviewed-by: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Commit-Queue: Michael van Ouwerkerk <mvanouwerkerk@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558110}