[omnibox][next] Fix crash when typing after @tabs mode.
https://crrev.com/c/7191448 changed the ordering of events such that it
is now possible for the selection to be out of bounds. That change
however fixed another crash so if we revert it then it will be
reintroduced. Ideally we would understand the issue more deeply and fix
the incorrect low however today is the last day for 144 beta merges and
this would fix the @tabs feature there.

Bug: b:468047546
Change-Id: I61398e7a5e20880e7eeac7da26a4aae4cc0217ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7252897
Auto-Submit: Aviv Kiss <avivkiss@google.com>
Reviewed-by: Justin Donnelly <jdonnelly@chromium.org>
Commit-Queue: Justin Donnelly <jdonnelly@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558122}