[context-input][Android] Disable separate request IDs for multi-context viewport images in Composebox.
This change sets
`use_separate_request_ids_for_multi_context_viewport_images` to false in
the `QueryControllerConfigParams` for Composebox. This matches the field
trial configs for NTP and omnibox for desktop.

While it would be possible to update the default value for the param in
QueryControllerConfigParams, I opted to make the change here in
ComposeBoxQueryControllerBridge because 1) updating the flag default
would require updating tests and 2) the server will actually eventually
be migrating to back to using separate request ids for viewport images.
This disablement will only be temporary.

Bug: b:466428336
Change-Id: Iaf11f79bde032cc3d7ddc29325f8233adf7c8321
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7254423
Auto-Submit: Jason Hu <hujasonx@google.com>
Commit-Queue: Patrick Noland <pnoland@chromium.org>
Reviewed-by: Patrick Noland <pnoland@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558145}