Create ShowImmersiveUI to add Reading Mode to the overlay view.
ReadAnythingController's ShowImmersiveUI method instantiates and
displays the immersive overlay, and passes the Reading Mode WebUI to the
overlay.

Introduces ReadAnythingImmersiveWebView, a webview to host the Reading
Mode WebUI within the ReadAnythingImmersiveOverlayView.

The ShowUI and CloseUI in the ReadAnythingImmersiveWebView are not
explicitly called in this code, because the WebUI calls those methods on
it (because it's the WebUI's embedder)

Having the immersive overlay cover the main webpage caused the main
webpage to not send Accessibility events because of optimizations the
render makes for occluded pages. This broke Reading Mode's distillation
triggers. This CL therefore uses WebContents::IncrementCapturerCount to
force the main contents to be treated as visible. This capture is
managed by CaptureMainContentsAsVisible and ReleaseMainContentsCapture,
called when the immersive UI becomes visible/hidden.

This CL is a no-op- no one currently calls ShowImmersiveUI other than
tests, because the ability to close the overlay isn't built yet and the
UI entrypoints aren't built yet either.

Bug: 450948683
Change-Id: Iee81eb7c9ba68fb255b8e94a7e9263dbfaff214a
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7213205
Commit-Queue: Eitan Goldberger <eitang@google.com>
Reviewed-by: Kristi Saney <kristislee@google.com>
Reviewed-by: Jacobo Misael Tapia <jatapiaro@google.com>
Cr-Commit-Position: refs/heads/main@{#1558196}