Fix BackgroundLoaderOffliner crash when terminating after completion callback
This change is a speculative fix for a crash in BackgroundLoaderOffliner
by ensuring that TerminateLoadIfInProgress does not attempt to call
Cancel if the completion callback has already been invoked and nulled
out. This could happen if TerminateLoadIfInProgress is called more than
once.

The ResetState call in CanDownload is also made synchronous to make sure
that pending_request_ is nullified when completion_callback_ runs. The
ResetState method should no longer need to be posted as a separate task,
because https://crrev.com/c/517812 handles any edge cases with stack
corruption.

Bug: 467089318
Change-Id: I651c07c6594cef2b1c6ebdcf5b6f2b73245fd137
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7253885
Reviewed-by: Dan Harrington <harringtond@chromium.org>
Commit-Queue: Aldo Culquicondor <acondor@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558255}