Remove use of BodyAsStringCallbackDeprecated (ash/ambient)
Updates an existing use of the version of
SimpleURLLoader::DownloadToString that takes a
BodyAsStringCallbackDeprecated to use the version that takes a
BodyAsStringCallback. This involves updating the callback function to
accept a std::optional<std::string> instead of a
std::unique_ptr<std::string>. This CL touches a usage in ash/ambient
that wasn't indexed in code search for some reason (I've opened
b/468023269 for that).

Bug: 40258809
Change-Id: I9a9cceb2a31596f5887559d125b7d0b9ab388426
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7254135
Reviewed-by: Xiaohui Chen <xiaohuic@chromium.org>
Commit-Queue: Andrew Williams <awillia@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558297}