Add semantic label plumbing
When the runtime knows what semantic labels should be present, send the
data over mojom so that the page can surface this data. There is a
very limited set of semantic labels available, and thus we intentionally
plumb this across via an enum so that we can be intentional about the
labels that we allow to be shared by the runtime.

Bug: 394636076
Change-Id: I5d07d87ece0dbe8e4346d9d10e751a0bb4f03516
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7247024
Reviewed-by: Elly FJ <ellyjones@chromium.org>
Commit-Queue: Alexander Cooper <alcooper@chromium.org>
Reviewed-by: Brandon Jones <bajones@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558301}