gfx::HDRMetadata: Remove gfx::HdrMetadataAgtm
This is towards moving gfx::HDRMetadata to be skhdr::Metadata, so that
we can share HDR rendering across Chrome and Android, and so we can
share metadata parsing between Chrome and Skia codecs.

skhdr::Metadata just stores an sk_sp<SkData> of the serialize metadata.
We do not parse the metadata until the very last moment, when we are
rendering the image/video that uses it (and when we do that, we should
have a cache of the parsed result and its shaders, etc).

Remove the Remove gfx::HdrMetadataAgtm structure, and make it just be a
(lonely) private member on gfx::HDRMetadata. We will do the same with
the remaining members (adding accessors to mirror the Skia structures),
and then we will eventually replace the structure with skhdr::Metadata.

Bug: 395659818
Change-Id: I8d0be6756d651ec46963c262c22983502a74e7a3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7245979
Commit-Queue: ccameron chromium <ccameron@chromium.org>
Reviewed-by: Dale Curtis <dalecurtis@chromium.org>
Reviewed-by: Daniel Cheng <dcheng@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558303}