[Extensions] Disallow NTP overrides for split mode extensions
Extensions were incorrectly able to override the new tab page if they
were set to "incognito": "split", but we state in the documentation that
extensions should never be able to override the new tab page.

This CL fixes this and updates the recently added tests to document
this.

Bug: 455164566
Fixed: 460732314
Change-Id: If2d956a1c9a1f3b6cec4bc8a5de247f5050750a6
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7139819
Reviewed-by: Devlin Cronin <rdevlin.cronin@chromium.org>
Commit-Queue: Tim <tjudkins@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558308}