[Extensions] Clean up ExtensionTabUtil::OpenTab()'s browser selection
ExtensionTabUtil::OpenTab() tries to find or create a browser to open
a tab in, but the browser selection is rather clunky and inconsistent.
This CL isolates more of the logic and improves a few egregious cases
while trying to minimize behavior changes and the risk of breakage for
any extensions relying on the existing behavior.

Some changes:
* Reduce and concentrate the number of calls to fall back to a tabbed
  browser or create a new browser.
* Prevent the possibility of creating *multiple* new browsers.
* Prevent some cases where we could create a new browser only to throw
  an error (and not open the tab).

In addition to addressing some bad scenarios, this will also make this
method easier to port to the experimental desktop android platform.

Bug: 468223125, 430344931
Change-Id: Ibd02d43ade81c8db46495d46edaf717c05d05a14
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7256997
Reviewed-by: James Cook <jamescook@chromium.org>
Commit-Queue: Devlin Cronin <rdevlin.cronin@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558367}