Add XR devices for AndroidDesktopWebPrefsLargeDisplays
This change makes XR devices have the same behavior as tablet for
AndroidDesktopWebPrefsLargeDisplays. A flickering bug was visible on
tablets and XR devices, where if the users hovered certain elements on
spotify, intense flickering was seen. This was resolved on tablet by
enabling this flag to get certain webprefs. We are doing the same for
XR.

Tested video:
https://drive.google.com/file/d/1E1GAI5Oy0HmozMrI-On3-Cuw8LvgVk-2/view?usp=sharing&resourcekey=0-QS58yKjvw62wGwbLVpJXcA

Bug: 453867322
Fixed: 453867322
Change-Id: I9fa8502a2424325e617776d96e627d74266985b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7249788
Reviewed-by: Dave Tapuska <dtapuska@chromium.org>
Commit-Queue: Nick Mondello <mondello@google.com>
Reviewed-by: Gurmeet Kalra <gurmeetk@google.com>
Cr-Commit-Position: refs/heads/main@{#1558373}