Preserve lazy loading behavior when an image is side-loaded
According to the spec, if the resource for an <img loading="lazy"> is
loaded via other means (e.g. fetch()), then the <img> should not
fire its load event until it enters the viewport.

Putting this behind a flag in case sites are relying on the legacy
broken behavior.

Bug: 375209498
Change-Id: I58f47ae1efdb953cdffc642eb68264484d1d20b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7240346
Commit-Queue: Stefan Zager <szager@chromium.org>
Reviewed-by: Philip Rogers <pdr@chromium.org>
Cr-Commit-Position: refs/heads/main@{#1558440}