[Desktop Zoom] Update default fieldtrial experiment
This CL updated the fieltrial config for the desktop zoom feature to
use a non-default value, which is the eventual launch candidate. The
fact that we are scaling content 9% in some cases is not meaningful to
include in render tests, and so we disable the feature in some cases.

Bug: 450281745
Change-Id: I057a6cef9364fc88ad99913a007ad7ec803772bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromium/src/+/7238453
Reviewed-by: Theresa Sullivan <twellington@chromium.org>
Commit-Queue: Mark Schillaci <mschillaci@google.com>
Cr-Commit-Position: refs/heads/main@{#1558444}